[internal] Rename files
