+++ src/client/components/category/CategoryBadge.component.tsx
